adds a "start" button
the start button is now what the user uses to start entering values in
the app. before pressing start the rest of the app is disabled.

after pressing start the user cannot return to the home screen, without
refreshing the page.

this prevents a bug where we create a params file after changing the
selected dataset