switches to generating manifest.json with js
manifest.json has two sections which change, a list of packages, and a
list of files to deploy.

the list of packages for the most part doesn't change often, but the
list of files causes lots of merge conflicts.

this adds a partial manifest.json file and adds a javascript file to
regenerate the files section without needing to set up R.