adds custom variants
- closes #472
- closes #494

for both Manchester and Milton Keynes, adds their custom population variants as the default options.
leaves all the other projections in place, but with a value of 0 by default.

if using existing scenarios, then they will continue to use the previously selected (probably principal) projection values and will need to be manually edited.

new scenarios will need the default params in the inputs selection app to be changed - will handle that separately.